extract colors and draw configs
