+++ drawings/colors.py
@@ -1,1 +1,17 @@
-Color = tuple[int, int, int]
+from enum import Enum
+
+class Color(Enum):
+    RED = (0, 0, 255)
+    GREEN = (0, 255, 0)
+    BLUE = (255, 0, 0)
+    WHITE = (255, 255, 255)
+    BLACK = (0, 0, 0)
+    YELLOW = (0, 255, 255)
+    ORANGE = (0, 165, 255)
+    PURPLE = (128, 0, 128)
+    PINK = (203, 192, 255)
+    BROWN = (42, 42, 165)
+    GRAY = (128, 128, 128)
+    CYAN = (255, 255, 0)
+    MAGENTA = (255, 0, 255)
+    TEAL = (128, 128, 0)
